Fix: squid:S1905, Redundant casts should not be used

+++ tomitribe-crest/src/test/java/org/tomitribe/crest/DefaultOptionListsAndOptionsTest.java
@@ -134,9 +134,9 @@ public class DefaultOptionListsAndOptionsTest extends TestCase {
             assertNotNull(list);
 
             final Iterator<Integer> it = list.iterator();
-            assertEquals((Integer) (int) 2, it.next());
-            assertEquals((Integer) (int) 3, it.next());
-            assertEquals((Integer) (int) 5, it.next());
+            assertEquals((Integer) 2, it.next());
+            assertEquals((Integer) 3, it.next());
+            assertEquals((Integer) 5, it.next());
             assertFalse(it.hasNext());
 
             assertEquals(Color.orange, colorParam);
